% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chardist.R
\name{chardist}
\alias{chardist}
\title{Generates distance matrix for categorical columns in two data sets.}
\usage{
chardist(dfa, dfb)
}
\arguments{
\item{dfa}{data set A to compare to data set B.}

\item{dfb}{data set B to compare to data set A.}
}
\value{
A distance matrix for the categorical variables between two data
sets.
}
\description{
Computes a distance matrix and distance (similarity) score for the categorical
columns in two data sets.
}
\examples{
chardist(penguins,penguins_raw)

}
