% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chardist_XY.R
\name{chardist_XY}
\alias{chardist_XY}
\title{Distance between two character variables}
\usage{
chardist_XY(varA, varB)
}
\arguments{
\item{varA}{vector from dataset A}

\item{varB}{vector from dataset B}
}
\description{
The distance between two character (or factor) variables is being calculated.
}
\examples{
# example code
chardist_XY(penguins$species, penguins_raw$Sex)
}
