% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charscore.R
\name{charscore}
\alias{charscore}
\title{Generates distance score for categorical columns.}
\usage{
charscore(dfa, dfb)
}
\arguments{
\item{dfa}{Data frame A to compare to `dfb`}

\item{dfb}{Data frame B to compare to `dfa`}
}
\value{
A list containing a dataframe with variable mappings & distances and
an overall dataframe distance score.
}
\description{
Computes a distance matrix and distance (similarity) score for the categorical
columns in two data sets.
}
\examples{
# example code

}
