% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadist.R
\name{datadist}
\alias{datadist}
\title{Difference Measure for Data Sets}
\usage{
datadist(dfa, dfb)
}
\arguments{
\item{dfa}{data.frame}

\item{dfb}{data.frame}
}
\value{
`numdist` is returned using the `wassersteinXY()` function to find a
distance matrix for every numerical variable combination between data frame A
and data frame B. `chardist` is found likewise.
}
\description{
Produces a numerical variable distance matrix and a categorical variable
distance matrix.
}
\examples{
# example code
dataA <- data.frame(a = c(1,5,3,2),
                    b = c(4,1,6,8),
                    c = c(3,6,1,7),
                    d = c("apple","orange","banana","mango"))
dataB <- data.frame(a = c(1,9,2,6),
                    b = c(4,2,6,9),
                    c = c(3,6,1,7),
                    d = c(3,7,9,2),
                    e = c("pear","grape","lemon","lime"))
}
