% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_cols.R
\name{date_cols}
\alias{date_cols}
\title{Adds or removes date columns from a data frame.}
\usage{
date_cols(dfA, dfB = NA, add = TRUE)
}
\arguments{
\item{dfA}{data frame to add date columns to or remove date columns from.}

\item{dfB}{optional second data frame to extract date columns from to be added
to dfA. Default is NA.}

\item{add}{binary operator to add or remove date column(s), where TRUE adds
the columns. Default is TRUE.}
}
\value{
A data frame with dates either added or removed, based on user
specification.
}
\description{
A function that will add date columns to a specified data frame
or remove them, based on user specification. Designed as a helper function
for numdist() and chardist().
}
\examples{
# example code


}
