% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscore.R
\name{dscore}
\alias{dscore}
\title{Generates distance score for a distance matrix.}
\usage{
dscore(dmat)
}
\arguments{
\item{dfa}{Data frame A to compare to `dfb`}

\item{dfb}{Data frame B to compare to `dfa`}
}
\value{
A list containing a dataframe with variable mappings & distances and
an overall dataframe distance score.
}
\description{
Computes a distance matrix and distance (similarity) score for the numerical
or categorical variables of two data frames.
}
\examples{
dscore(numdist(dataA, dataB))

}
