% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kgl_diamonds}
\alias{kgl_diamonds}
\title{List of five diamonds datasets from kaggle}
\format{
An object of class \code{list} of length 5.
}
\usage{
kgl_diamonds
}
\description{
The `diamonds` dataset is well-known as part of the ggplot2 package.
The kaggle platform hosts an astonishing 176 datasets named `diamonds` ([search result](https://www.kaggle.com/search?q=diamonds+in%3Adatasets)
from 2026, Apr 26). Five of these datasets are included in the
`kgl_diamonds` list. Each item of the list contains one dataset, named
after their respective kaggle author.
}
\keyword{datasets}
