% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins_sample}
\alias{penguins_sample}
\title{Sample of the `penguins` dataset}
\format{
An object of class \code{data.frame} with 200 rows and 8 columns.
}
\usage{
penguins_sample
}
\description{
The `penguins` dataset is well-known as part of the datasets package.
`penguins_sample` is a dataset containing a random sample 200 observations
from the original `penguins` dataset.
The purpose of this dataset is to demonstrate the use of `datadist` functions
on datasets that have been sampled from the original dataset.
}
\keyword{datasets}
