% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins_whodunit}
\alias{penguins_whodunit}
\title{Mutation of the `penguins` dataset}
\format{
An object of class \code{data.frame} with 344 rows and 8 columns.
}
\usage{
penguins_whodunit
}
\description{
The `penguins` dataset is well-known as part of the datasets package.
`penguins_whodunit` is a dataset that has mutated the original `penguins`
dataset in various manners.
The purpose of this dataset is to demonstrate the use of `datadist` functions
on datasets that are a mutated version of their original dataset.
}
\keyword{datasets}
