% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_single_cat.R
\name{rm_single_cat}
\alias{rm_single_cat}
\title{Removes categorical variable columns with only one category.}
\usage{
rm_single_cat(df)
}
\arguments{
\item{df}{data frame to remove one category categorical variables from.}
}
\value{
A data frame with categorical variables containing only one category
removed.
}
\description{
A function that removes categorical variables from a data frame
if they only contain one category. Desiged for use with chardist().
}
\examples{
# example code
head(rm_single_cat(penguins_raw))

}
