% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wassersteinXY.R
\name{wassersteinXY}
\alias{wassersteinXY}
\title{Calculate the Wasserstein distance between two random variables}
\usage{
wassersteinXY(X, Y, p = 1, knots = max(c(length(X), length(Y))))
}
\arguments{
\item{X}{a real-valued random variable}

\item{Y}{a real-valued random variable}

\item{p}{non-negative value identifying the norm. $p$ = 1 is Manhattan norm, $p=2$ is Euclidean distance.}

\item{knots}{non negative integer value specifying the number of evaluations within the ranges of X and Y.}
}
\value{
single positive value. Larger values indicate less similarity.
}
\description{
The Wasserstein distance between two numeric variables X and Y is defined
as the distance between their distribution functions $F_X$ and $F_Y$ measured
in the p-norm.
}
\examples{

x <- runif(1000)
y <- runif(500)
wassersteinXY(x, y, knots=50)
transport::wasserstein1d(x, y)
}
