% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chardist.R
\name{chardist}
\alias{chardist}
\title{Generates distance matrix for categorical columns in two data sets.}
\usage{
chardist(dfa, dfb)
}
\arguments{
\item{dfa}{Data frame A to compare to `dfb`}

\item{dfb}{Data frame B to compare to `dfa`}
}
\value{
A distance matrix for the categorical variables between two data sets.
}
\description{
Computes a distance matrix and distance (similarity) score for the categorical
columns in two data sets.
}
\examples{
chardist(penguins[, c("species", "island", "sex")],
         penguins_raw[, c("Species", "Region", "Island", "Sex")])

}
