% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chardist.R
\name{chardist}
\alias{chardist}
\title{Generates distance matrix for categorical columns in two data sets.}
\usage{
chardist(dfa, dfb)
}
\arguments{
\item{dfa}{data set A or variable A to compare to data set B or variable
B (`dfb`).}

\item{dfb}{data set B  or variable B to compare to data set A or variable
A (`dfa`).}
}
\value{
A distance matrix for the categorical variables between two data
sets, plus the distance score.
}
\description{
Computes a distance matrix and distance (similarity) score for the categorical
columns in two data sets. Can compute a distance matrix between the categories
of two individual categorical variables.
}
\examples{
<<<<<<< HEAD
chardist(penguins,penguins_raw)
=======
chardist(penguins[, c("species", "island", "sex")],
         penguins_raw[, c("Species", "Region", "Island", "Sex")])
>>>>>>> f4f95be14e43e87ace4285dfd6b92ecc6e83bddc

}
