% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chardist_XY.R
\name{chardist_XY}
\alias{chardist_XY}
\title{Compares two catecorical (or factor) columns between two data sets.}
\usage{
chardist_XY(varA, varB)
}
\arguments{
\item{varA}{vector from dataset A}

\item{varB}{vector from dataset B}
}
\value{
A a binary response indicating if two variables are identical, a
marginal distance between the two variables as factors, and objects
returned by use of dscore().
}
\description{
The distance between two character (or factor) variables is being calculated.
}
\examples{
# example code
chardist_XY(penguins$species, penguins_raw$Sex)

}
