% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{datadist}
\alias{datadist}
\title{Difference Measure for Data Sets}
\usage{
datadist(a, b)
}
\arguments{
\item{a}{data.frame}

\item{b}{data.frame}
}
\value{
A list of objects.
}
\description{
Test function for data distance function... more to come :)
}
\examples{
# example code
dataA <- data.frame(a = c(1,5,3,2),
                    b = c(4,1,6,8),
                    c = c(3,6,1,7),
                    d = c("apple","orange","banana","mango"))
dataB <- data.frame(a = c(1,9,2,6),
                    b = c(4,2,6,9),
                    c = c(3,6,1,7),
                    d = c(3,7,9,2),
                    e = c("pear","grape","lemon","lime"))
datadist(dataA, dataB)
}
