% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numdist.R
\name{numdist}
\alias{numdist}
\title{Difference Measure for Data Sets}
\usage{
numdist(dfa, dfb)
}
\arguments{
\item{dfa}{data.frame}

\item{dfb}{data.frame}
}
\value{
`numdist` is returned using the `wassersteinXY()` function to compute a
distance matrix for every numerical variable combination between data frame A
and data frame B.
}
\description{
Produces a numerical variable distance matrix between two data sets.
}
\examples{
# example code
dataA <- data.frame(a = c(1,5,3,2),
                    b = c(4,1,6,8),
                    c = c(3,6,1,7),
                    d = c("apple","orange","banana","mango"))
dataB <- data.frame(a = c(1,9,2,6),
                    b = c(4,2,6,9),
                    c = c(3,6,1,7),
                    d = c(3,7,9,2),
                    e = c("pear","grape","lemon","lime"))
numdist(dataA, dataB)
}
