% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numscore.R
\name{numscore}
\alias{numscore}
\title{Generates distance score}
\usage{
numscore(dmat)
}
\arguments{
\item{dmat}{Distance matrix derived from use of `datadist()`}
}
\value{
A list containing a dataframe with variable mappings & distances and
an overall dataframe distance score.
}
\description{
Computes a distance score for two matrices after use of `datadist()`
}
\examples{
# example code
}
