% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numscore.R
\name{numscore}
\alias{numscore}
\title{Generates distance score for numerical columns.}
\usage{
numscore(dmat)
}
\arguments{
\item{dmat}{matrix of distances between pairwise numeric variables calculated in datadist}
}
\value{
A list containing a dataframe with variable mappings & distances and
an overall dataframe distance score.
}
\description{
Computes a distance matrix and distance (similarity) score for the numerical
variables of two data frames.
}
\examples{
# example code

}
